requirements.txt
# platform: linux-64
datasets=2.0.0
dgl-cu113=0.8.2.post1
networkx=2.7.1=pyhd8ed1ab_0
numpy=1.21.1=py38h9894fe3_0
pandas=1.2.3=py38ha9443f7_0
transformers==4.17.0
tokenizers==0.12.1
pytorch=1.8.0
scikit-learn=0.24.1=py38ha9443f7_0
scipy=1.6.2=py38h91f5cce_0
